@@ -10,21 +10,21 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4.0"
     },
     "require-dev": {
         "mockery/mockery": "0.7.2",
-        "illuminate/view": "4.0.*",
+        "illuminate/view": "4.2.*",
         "phpunit/phpunit": "3.7.*"
     },
     "autoload": {
-        "psr-0": {
-            "Robbo\\Presenter": "src/"
+        "psr-4": {
+            "Robbo\\Presenter\\": "src/"
         }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.2-dev"
+            "dev-master": "1.3-dev"
         }
     },
     "minimum-stability": "stable"
